MiddlewareApi/requirements.txt
autopep8==1.4.4
Django==2.2.5
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
mysqlclient==1.4.4
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
